app/requirements-pip3.txt
pandas==1.3.5
flask==2.3.2
psycopg2-binary==2.9.6
click==8.1.6
